// src/main/clients/homeassistant/enums/HaServiceFieldType.ts
export enum HaServiceFieldType {
    NONE = 'none',
    NUMBER = 'number',
    TEXT = 'text',
    BOOLEAN = 'boolean',
    SELECT = 'select',
    OBJECT = 'object',
    COLOR_RGB = 'color_rgb',
    COLOR_TEMP = 'color_temp',
    CONSTANT = 'constant',
    DATE = 'date',
    TIME = 'time',
    DATETIME = 'datetime',
    ENTITY = 'entity',
    DEVICE = 'device',
    THEME = 'theme'
}