// src/main/clients/homeassistant/enums/HomeassistantStatus.ts
export enum HomeassistantStatus {
    BOOTING = 'Booting',
    NO_SETTINGS = '',
    NO_URL_AND_TOKEN = 'No URL and Token',
    NO_URL = 'No URL',
    NO_TOKEN = 'No Token',
    UNAUTHORIZED = 'Unauthorized',
    NO_CONNECTION = 'No Connection',
    CONNECTED = 'Connected',
    CONNECTING = 'Connecting'
}